add timer
