+++ UIKit/Lesson 11/Lesson 11/AppDelegate.swift
@@ -13,8 +13,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-//        UITabBar.appearance().barTintColor = .black
-//        UITabBar.appearance().tintColor = .red
+        UITabBar.appearance().barTintColor = .black
+        UITabBar.appearance().tintColor = .orange
         return true
     }
