+++ UIKit/Lesson 11/Lesson 11/GlobalTimeViewController.swift
@@ -1,0 +1,29 @@
+//
+//  GlobalTimeViewController.swift
+//  Lesson 11
+//
+//  Created by T1GER on 02.07.2021.
+//
+
+import UIKit
+
+class GlobalTimeViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
